rohith/requirements.txt
# --- Core Project Ivy Dependencies ---
langgraph>=0.0.26
langchain-openai>=0.0.5
langchain-google-genai>=0.0.10
langchain-groq>=0.0.1
streamlit>=1.30.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pandas<3.0.0
numpy>=1.24.0
yfinance>=0.2.36
python-dotenv>=1.0.0
simple-salesforce>=1.12.0

# --- V2 Upgrade Dependencies ---
# UI and Visualization (Phase 7)
plotly==5.18.0

# PDF Generation (Phase 6)
weasyprint==60.0

# Modeling & Machine Learning (Phase 5)
transformers>=4.35.0
torch
scikit-learn>=1.3.0
joblib>=1.3.2

# Reliability & Error Handling (Phase 4)
sentry-sdk==1.40.0
tenacity==8.2.3
markdown

# Testing Infrastructure (Phase 8)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0